chore(cms-api): move crud input embeddable tests into own file
